requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
# Exporter for Google Cloud
opentelemetry-exporter-gcp-trace
